requirements.txt
pandas>=2.1
pyyaml>=6.0
python-dotenv>=1.0
